+++ FirebaseStorage/Sources/AsyncAwait.swift
@@ -110,6 +110,42 @@ public extension StorageReference {
     }
   }
 
+  /// Asynchronously uploads from a FileHandle to the currently specified StorageReference.
+  ///
+  /// - Parameters:
+  ///   - fileHandle: A file handle to the data to upload.
+  ///   - metadata: Optional StorageMetadata containing additional information (MIME type, etc.)
+  ///              about the object being uploaded.
+  ///   - onProgress: An optional closure function to return a `Progress` instance while the
+  /// upload proceeds.
+  /// - Throws: An error if the operation failed.
+  /// - Returns: `StorageMetadata` with additional information about the object being uploaded.
+  func putFileHandleAsync(_ fileHandle: FileHandle,
+                          metadata: StorageMetadata? = nil,
+                          onProgress: ((Progress?) -> Void)? = nil) async throws -> StorageMetadata {
+    guard let onProgress = onProgress else {
+      return try await withCheckedThrowingContinuation { continuation in
+        self.putFileHandle(fileHandle, metadata: metadata) { result in
+          continuation.resume(with: result)
+        }
+      }
+    }
+    let uploadTask = putFileHandle(fileHandle, metadata: metadata)
+    return try await withCheckedThrowingContinuation { continuation in
+      uploadTask.observe(.progress) {
+        onProgress($0.progress)
+      }
+      uploadTask.observe(.success) { _ in
+        continuation.resume(with: .success(uploadTask.metadata!))
+      }
+      uploadTask.observe(.failure) { snapshot in
+        continuation.resume(with: .failure(
+          snapshot.error ?? StorageError.internalError("Internal Storage Error in putFileHandleAsync")
+        ))
+      }
+    }
+  }
+
   /// Asynchronously downloads the object at the current path to a specified system filepath.
   ///
   /// - Parameters: