+++ FirebaseStorage/Sources/StorageReference.swift
@@ -185,6 +185,82 @@ import Foundation
     return task
   }
 
+  /// Asynchronously uploads from a FileHandle to the currently specified `StorageReference`,
+  /// without additional metadata.
+  /// - Parameters:
+  ///   - fileHandle: A file handle to the data to upload.
+  ///   - metadata: `StorageMetadata` containing additional information (MIME type, etc.)
+  ///       about the object being uploaded.
+  /// - Returns: An instance of `StorageUploadTask`, which can be used to monitor or manage the
+  /// upload.
+  @objc(putFileHandle:metadata:)
+  @discardableResult
+  open func putFileHandle(_ fileHandle: FileHandle, metadata: StorageMetadata? = nil) -> StorageUploadTask {
+    return putFileHandle(fileHandle, metadata: metadata, completion: nil)
+  }
+
+  /// Asynchronously uploads from a FileHandle to the currently specified `StorageReference`.
+  /// - Parameter fileHandle A file handle to the data to upload.
+  /// - Returns: An instance of `StorageUploadTask`, which can be used to monitor or manage the
+  /// upload.
+  @objc(putFileHandle:) @discardableResult open func __putFileHandle(_ fileHandle: FileHandle) -> StorageUploadTask {
+    return putFileHandle(fileHandle, metadata: nil, completion: nil)
+  }
+
+  /// Asynchronously uploads from a FileHandle to the currently specified `StorageReference`,
+  /// - Parameters:
+  ///   - fileHandle: A file handle to the data to upload.
+  ///   - metadata: `StorageMetadata` containing additional information (MIME type, etc.)
+  ///       about the object being uploaded.
+  ///   - completion: A closure that either returns the object metadata on success,
+  ///       or an error on failure.
+  /// - Returns: An instance of `StorageUploadTask`, which can be used to monitor or manage the
+  /// upload.
+  @objc(putFileHandle:metadata:completion:) @discardableResult
+  open func putFileHandle(_ fileHandle: FileHandle,
+                          metadata: StorageMetadata? = nil,
+                          completion: ((_: StorageMetadata?, _: Error?) -> Void)?) -> StorageUploadTask {
+    let putMetadata = metadata ?? StorageMetadata()
+    if let path = path.object {
+      putMetadata.path = path
+      putMetadata.name = (path as NSString).lastPathComponent as String
+    }
+    let task = StorageUploadTask(reference: self,
+                                 service: storage.fetcherServiceForApp,
+                                 queue: storage.dispatchQueue,
+                                 fileHandle: fileHandle,
+                                 metadata: putMetadata)
+    startAndObserveUploadTask(task: task, completion: completion)
+    return task
+  }
+
+
+/*
+  open func putFileHandle(_ fileHandle: FileHandle, metadata: StorageMetadata? = nil) -> StorageUploadTask {
+    return putFileHandle(fileHandle, metadata: metadata, completion: nil)
+  }
+
+
+    open func putFileHandle(_ fileHandle: FileHandle,
+                            metadata: StorageMetadata? = nil,
+                            completion: ((_: StorageMetadata?, _: Error?) -> Void)?) -> StorageUploadTask {
+        print("called this one")
+        let putMetadata: StorageMetadata = metadata ?? StorageMetadata()
+
+        if let path = path.object {
+          putMetadata.path = path
+          putMetadata.name = (path as NSString).lastPathComponent as String
+        }
+
+        let task = StorageUploadTask(reference: self,
+                                     service: storage.fetcherServiceForApp,
+                                     queue: storage.dispatchQueue,
+                                     fileHandle: fileHandle,
+                                     metadata: putMetadata)
+        startAndObserveUploadTask(task: task, completion: completion)
+        return task
+    }
+*/
   // MARK: - Downloads
 
   /// Asynchronously downloads the object at the `StorageReference` to a `Data` instance in memory.