+++ FirebaseStorage/Sources/StorageUploadTask.swift
@@ -87,6 +87,9 @@ import Foundation
       } else if let fileURL = self.fileURL {
         uploadFetcher.uploadFileURL = fileURL
         uploadFetcher.comment = "File UploadTask"
+      } else if let fileHandle = self.uploadFileHandle {
+        uploadFetcher.uploadFileHandle = fileHandle
+        uploadFetcher.comment = "FileHandle UploadTask"
       }
       uploadFetcher.maxRetryInterval = self.reference.storage.maxUploadRetryInterval
 
@@ -194,6 +197,7 @@ import Foundation
   private var fetcherCompletion: ((Data?, NSError?) -> Void)?
   private var uploadMetadata: StorageMetadata
   private var uploadData: Data?
+  private var uploadFileHandle: FileHandle?
   // Hold completion in object to force it to be retained until completion block is called.
   var completionMetadata: ((StorageMetadata?, Error?) -> Void)?
 
@@ -204,9 +208,11 @@ import Foundation
        queue: DispatchQueue,
        file: URL? = nil,
        data: Data? = nil,
+       fileHandle: FileHandle? = nil,
        metadata: StorageMetadata) {
     uploadMetadata = metadata
     uploadData = data
+    uploadFileHandle = fileHandle
     super.init(reference: reference, service: service, queue: queue, file: file)
 
     if uploadMetadata.contentType == nil {
@@ -220,6 +226,9 @@ import Foundation
 
   private func contentUploadError() -> NSError? {
     if uploadData != nil {
+      return nil
+    }
+    if uploadFileHandle != nil {
       return nil
     }
     if let resourceValues = try? fileURL?.resourceValues(forKeys: [.isRegularFileKey]),