+++ FirebaseStorage/Tests/Unit/StorageAPITests.swift
@@ -67,6 +67,14 @@ final class StorageAPITests: XCTestCase {
     _ = ref.putFile(from: file, metadata: metadata) { (result: StorageMetadata?, error: Error?) in
     }
 
+    let fileHandle = FileHandle.nullDevice
+    _ = ref.putFileHandle(fileHandle)
+    _ = ref.putFileHandle(fileHandle, metadata: metadata)
+    _ = ref.putFileHandle(fileHandle, metadata: metadata) { (result: Result<StorageMetadata, Error>) in
+    }
+    _ = ref.putFileHandle(fileHandle, metadata: metadata) { (result: StorageMetadata?, error: Error?) in
+    }
+
     _ = ref.getData(maxSize: 122) { (result: Data?, error: Error?) in
     }
     _ = ref.getData(maxSize: 122) { (result: Result<Data, Error>) in
@@ -85,7 +93,7 @@ final class StorageAPITests: XCTestCase {
 
     ref.listAll { (listResult: StorageListResult?, error: Error?) in
     }
-    ref.listAll { (result: Result<StorageListResult, Error>) in
+    ref.listAll { (result: Result<StorageListResult, Error>) inÎ
     }
 
     ref.list(maxResults: 123) { (listResult: StorageListResult?, error: Error?) in